requirements.txt
flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.3.2
werkzeug==2.3.7
reportlab==4.0.4
arabic-reshaper==3.0.0
python-bidi==0.4.2
pillow
flask-login==0.6.3
pymongo==4.6.1
bcrypt==4.1.2
flask-wtf==1.2.1
pyOpenSSL==24.0.0
asgiref==3.7.2
hypercorn==0.15.0